@@ -7,30 +7,36 @@ import 'package:transparent_image/transparent_image.dart';
 import 'gif_page.dart';
 
 class HomePage extends StatefulWidget {
+  const HomePage({super.key});
+
   @override
-  _HomePageState createState() => _HomePageState();
+  State<HomePage> createState() => _HomePageState();
 }
 
 class _HomePageState extends State<HomePage> {
-  String _search;
+  static const limit = 19;
+
+  late Future<Map> gifsFuture;
+
+  String _search = '';
   int _offset = 0;
 
   Future<Map> _getGifs() async {
-    http.Response response;
-    if (_search == null) {
-      response = await http.get(
-        'https://api.giphy.com/v1/gifs/trending?api_key=IyO7FLT2n9WFb7wJA4qx1cXf68IoBq42&limit=20&rating=g');
-    } else {
-      response = await http.get(
-        'https://api.giphy.com/v1/gifs/search?api_key=IyO7FLT2n9WFb7wJA4qx1cXf68IoBq42&q=$_search&limit=19&offset=$_offset&rating=g&lang=pt');
+    if (_search.isEmpty) {
+      return {};
     }
+
+    http.Response response;
+    response = await http.get(Uri.parse(
+        'https://api.giphy.com/v1/gifs/search?api_key=IyO7FLT2n9WFb7wJA4qx1cXf68IoBq42&q=$_search&limit=$limit&offset=$_offset&rating=g&lang=pt'));
+
     return json.decode(response.body);
   }
 
   @override
   void initState() {
     super.initState();
-    _getGifs().then(print);
+    gifsFuture = _getGifs();
   }
 
   @override
@@ -40,35 +46,39 @@ class _HomePageState extends State<HomePage> {
         backgroundColor: Colors.black12,
         centerTitle: true,
         title: Image.network(
-        'https://developers.giphy.com/static/img/dev-logo-lg.7404c00322a8.gif'),
+          'https://developers.giphy.com/branch/master/static/header-logo-0fec0225d189bc0eae27dac3e3770582.gif',
+        ),
       ),
       backgroundColor: Colors.black38,
       body: Column(
         children: <Widget>[
           Padding(
-            padding: const EdgeInsets.all(10),
+            padding: const EdgeInsets.all(8),
             child: TextField(
-              decoration: InputDecoration(
+              decoration: const InputDecoration(
                 labelText: 'Pesquise seu gif:',
                 labelStyle: TextStyle(color: Colors.white),
                 border: OutlineInputBorder(),
               ),
-              style: TextStyle(
+              style: const TextStyle(
                 color: Colors.white,
                 fontSize: 20,
-                fontWeight: FontWeight.bold),
+                fontWeight: FontWeight.bold,
+              ),
               textAlign: TextAlign.center,
               onSubmitted: (text) {
                 setState(() {
                   _search = text;
                   _offset = 0;
                 });
+
+                gifsFuture = _getGifs();
               },
             ),
           ),
           Expanded(
             child: FutureBuilder(
-              future: _getGifs(),
+              future: gifsFuture,
               builder: (context, snapshot) {
                 switch (snapshot.connectionState) {
                   case ConnectionState.waiting:
@@ -77,81 +87,104 @@ class _HomePageState extends State<HomePage> {
                       width: 200,
                       height: 200,
                       alignment: Alignment.center,
-                      child: CircularProgressIndicator(
+                      child: const CircularProgressIndicator(
                         valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
                         strokeWidth: 5.0,
                       ),
                     );
-                    default:
-                      if (snapshot.hasError) {
-                        return Container();
-                      } else {
-                        return _createGifTable(context, snapshot);
+                  default:
+                    debugPrint('TESTE: ${snapshot.data}');
+                    if (snapshot.hasError || snapshot.data?['data'] == null) {
+                      return const SizedBox();
+                    } else {
+                      final data = snapshot.data!['data'];
+
+                      if (data.isEmpty) {
+                        return const Center(
+                          child: Text(
+                            'Nenhum gif encontrado!',
+                            style: TextStyle(
+                              color: Colors.white,
+                              fontSize: 20.0,
+                            ),
+                          ),
+                        );
                       }
+
+                      return GridView.builder(
+                        padding: const EdgeInsets.all(8.0),
+                        gridDelegate:
+                            const SliverGridDelegateWithFixedCrossAxisCount(
+                          crossAxisCount: 2,
+                          crossAxisSpacing: 8.0,
+                          mainAxisSpacing: 8.0,
+                        ),
+                        itemCount: data.length + 1,
+                        itemBuilder: (context, index) {
+                          if (index == data.length) {
+                            return GestureDetector(
+                              child: const Column(
+                                mainAxisAlignment: MainAxisAlignment.center,
+                                children: <Widget>[
+                                  Icon(
+                                    Icons.add,
+                                    color: Colors.white,
+                                    size: 60.0,
+                                  ),
+                                  Text(
+                                    'Carregar mais...',
+                                    style: TextStyle(
+                                      color: Colors.white,
+                                      fontSize: 20.0,
+                                    ),
+                                  ),
+                                ],
+                              ),
+                              onTap: () {
+                                if (data.length < limit) {
+                                  return;
+                                }
+
+                                setState(() {
+                                  _offset += limit;
+                                });
+
+                                gifsFuture = _getGifs();
+                              },
+                            );
+                          }
+
+                          final image =
+                              data[index]['images']['fixed_height']['url'];
+
+                          return GestureDetector(
+                            child: FadeInImage.memoryNetwork(
+                              placeholder: kTransparentImage,
+                              image: image,
+                              height: 300.0,
+                              fit: BoxFit.cover,
+                            ),
+                            onTap: () {
+                              Navigator.push(
+                                context,
+                                MaterialPageRoute(
+                                  builder: (context) => GifPage(
+                                    gifData: data[index],
+                                  ),
+                                ),
+                              );
+                            },
+                            onLongPress: () => Share.share(image),
+                          );
+                        },
+                      );
+                    }
                 }
-              }
+              },
             ),
           ),
         ],
       ),
-    );
-  }
-
-  int _getCount(List data) {
-    if (_search == null || _search.isEmpty) {
-      return data.length;
-    } else {
-      return data.length + 1;
-    }
-  }
-
-  Widget _createGifTable(BuildContext context, AsyncSnapshot snapshot) {
-    return GridView.builder(
-      padding: EdgeInsets.all(10.0),
-      gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-        crossAxisCount: 2,
-        crossAxisSpacing: 10.0,
-        mainAxisSpacing: 10.0,
-      ),
-      itemCount: _getCount(snapshot.data['data']),
-      itemBuilder: (context, index) {
-        if (_search == null || index < snapshot.data['data'].length) {
-          return GestureDetector(
-            child: FadeInImage.memoryNetwork(
-              placeholder: kTransparentImage,
-              image: snapshot.data
-                ['data'][index]['images']['fixed_height']['url'],
-              height: 300.0,
-              fit: BoxFit.cover),
-              onTap: () {
-                Navigator.push(context, MaterialPageRoute(
-                  builder: (context) => GifPage(snapshot.data['data'][index])));
-              },
-              onLongPress: () {
-                Share.share(snapshot.data['data'][index]['images']
-                  ['fixed_height']['url']);
-              },
-            );
-        } else {
-          return Container(
-            child: GestureDetector(
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  Icon(Icons.add, color: Colors.white, size: 60.0),
-                  Text('Carregar mais...',
-                    style: TextStyle(color: Colors.white, fontSize: 20.0))
-                ],
-              ),
-              onTap: () {
-                setState(() {
-                  _offset += 19;
-                });
-              },
-            ),
-          );
-        }
-      }
     );
   }
 }
