@@ -2,27 +2,31 @@ import 'package:flutter/material.dart';
 import 'package:share/share.dart';
 
 class GifPage extends StatelessWidget {
-  final Map _gifData;
-  GifPage(this._gifData);
+  const GifPage({
+    super.key,
+    required this.gifData,
+  });
+
+  final Map gifData;
 
   @override
   Widget build(BuildContext context) {
+    final image = gifData['images']['fixed_height']['url'];
+
     return Scaffold(
       appBar: AppBar(
-        title: Text(_gifData['title']),
+        title: Text(gifData['title']),
         backgroundColor: Colors.white,
         actions: <Widget>[
           IconButton(
-            icon: Icon(Icons.share),
-            onPressed: (){
-              Share.share(_gifData['images']['fixed_height']['url']);
-            },
+            icon: const Icon(Icons.share),
+            onPressed: () => Share.share(image),
           )
         ],
       ),
       backgroundColor: Colors.black,
       body: Center(
-        child: Image.network(_gifData['images']['fixed_height']['url']),
+        child: Image.network(image),
       ),
     );
   }
